update parent pom version to 8-snapshot and include Davids code inside random package
Current status:

Random 200 and 500 codes:
http://localhost:8080/twitter/api/followers/ids

Random Spike and normal response time (not working as expected):
http://localhost:8080/twitter/api/followers/list
