+++ pom.xml
@@ -14,7 +14,7 @@
   <parent>
     <groupId>com.tomitribe</groupId>
     <artifactId>tomitribe-parent</artifactId>
-    <version>7</version>
+    <version>8-SNAPSHOT</version>
   </parent>
 
   <!-- Change: groupId, artifactId, version -->
@@ -78,8 +78,8 @@
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-compiler-plugin</artifactId>
         <configuration>
-          <source>1.7</source>
-          <target>1.7</target>
+          <source>1.8</source>
+          <target>1.8</target>
         </configuration>
       </plugin>
     </plugins>
