+++ src/main/java/com/twitter/dev/random/Users.java
@@ -1,10 +1,13 @@
-/* =====================================================================
+/*
+ * Tomitribe Confidential
  *
- * Copyright (c) 2011 David Blevins.  All rights reserved.
+ * Copyright Tomitribe Corporation. 2014
  *
- * =====================================================================
+ * The source code for this program is not published or otherwise divested
+ * of its trade secrets, irrespective of what has been deposited with the
+ * U.S. Copyright Office.
  */
-package com.twitter;
+package com.twitter.dev.random;
 
 import java.util.ArrayList;
 import java.util.List;
